+++ CHANGES.rst
@@ -5,6 +5,7 @@ tip (unreleased)
 ----------------
 - Removed some incompatibilities with non-default admin sites (gh-92)
 - Fixed error caused by ``HistoryRequestMiddleware`` during anonymous requests (gh-115 fixes gh-114)
+- Added workaround for clashing related historical accessors on User (gh-121)
 
 1.5.0 (2014-08-17)
 ------------------